Move util methods outside of unit test file
